Refactored downloader and redid playlist file mgnt
The downloader code is now based on a class instance instead of a module
function.  The the default extract_info function is now wrapped for
error management.

File downloading and management has also been changed.  Files are now
directly downloaded to the audio_cache folder, and only generic
extractor files need to be renamed with the hash.  This fixes stupid
issues with youtube changing encoder versions or whatever the fuck they
do.